Univariate Mapping Fully Works
Saving univariate maps now possible, due to a re-write of how data is
handled (previously, an object was destroyed before a pointer to it was
referenced, which made interacting with the color map result in the
program crashing).  QCPColorMap data is now initialized inside the
MapView class constructor, rather than SpecMap::Univariate.